Add information about absolute paths
While I was using the documentation, I was stuck trying to get `override.styl` to import font files. 

A user on the forum.vuejs.org website helped me solve this one. 

Should be useful for others to know. 

Not sure of all details regarding when to use absolute paths, and when to not use them.